requirements.txt
cachelib==0.9.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
Flask==2.2.5
Flask-Caching==2.1.0
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
isort==5.11.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
pluggy==1.2.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytest==7.4.2
python-dotenv==0.21.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
tomli==2.0.1
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.6
Werkzeug==2.2.3
zipp==3.15.0
